Updated /contracts
Added another file to adjust the NFT so it mints for free, instead of requiring 10 ERC-20 tokens.

Also, added a test and updated the deploy script and added a simple staking smart contract that gives ten of some ERC-20 tokens per day.

Maybe we can enhance the staking mechanism, however I will start working on a frontend that displays a free mint.